moved examples in apps/api
