+++ apps/api/src/endpoints/endpoints.module.ts
@@ -3,12 +3,14 @@ import { DynamicModuleUtils } from "@libs/common";
 import { AuthModule } from "./auth/auth.module";
 import { TokenModule } from "./token/token.module";
 import { UserModule } from "./user/user.module";
+import { ExampleModule } from "./example/example.module";
 
 @Module({
   imports: [
     AuthModule,
     TokenModule,
     UserModule,
+    ExampleModule,
   ],
   providers: [
     DynamicModuleUtils.getNestJsApiConfigService(),
