--- libs/common/src/example/example.module.ts
@@ -1,23 +1,0 @@
-import { Module } from "@nestjs/common";
-import { ExampleService } from "./example.service";
-import { DynamicModuleUtils } from "../utils";
-import { CommonConfigModule } from "../config";
-
-@Module({})
-export class ExampleModule {
-  static forRoot() {
-    return {
-      module: ExampleModule,
-      imports: [
-        CommonConfigModule,
-        DynamicModuleUtils.getCachingModule(),
-      ],
-      providers: [
-        ExampleService,
-      ],
-      exports: [
-        ExampleService,
-      ],
-    };
-  }
-}
