--- libs/common/src/example/index.ts
@@ -1,4 +1,0 @@
-export * from './example.service';
-export * from './example.controller';
-export * from './example.module';
-export * from './entities';
