+++ libs/common/src/index.ts
@@ -3,5 +3,4 @@ export * from './metrics';
 export * from './pubsub';
 export * from './config';
 export * from './utils';
-export * from './example';
 export * from './health-check';
