+++ libs/entities/src/entities/example.filter.ts
