+++ libs/entities/src/entities/example.ts
