+++ libs/entities/src/index.ts
@@ -1,1 +1,2 @@
+export * from './entities';
 export * from './schemas';
