+++ libs/services/src/example/index.ts
@@ -1,0 +1,1 @@
+export * from './example.service';
