+++ libs/services/src/index.ts
@@ -1,2 +1,3 @@
+export * from './example';
 export * from './token';
 export * from './user';
